requirements.txt
flask
pillow
redis
rq